feat: add toggle-teleport command to VanishCommand and bump version to 1.21.11-1.0.10-SNAPSHOT
