+++ gradle.properties
@@ -1,4 +1,4 @@
 kotlin.code.style=official
 kotlin.stdlib.default.dependency=false
 org.gradle.parallel=true
-version=1.21.11-1.0.9-SNAPSHOT
+version=1.21.11-1.0.10-SNAPSHOT
