--- surf-vanish-paper/src/main/kotlin/dev/slne/surf/vanish/paper/command/subcommands/VanishTeleportHelperCommand.kt
@@ -1,22 +1,0 @@
-package dev.slne.surf.vanish.paper.command.subcommands
-
-import dev.jorel.commandapi.kotlindsl.playerExecutor
-import dev.jorel.commandapi.kotlindsl.subcommand
-import dev.slne.surf.surfapi.core.api.messages.adventure.sendText
-import dev.slne.surf.vanish.paper.listener.HotkeyListener
-
-fun teleportHelperCommand() = subcommand("teleport-helper") {
-
-    playerExecutor { player, _ ->
-        val newState = HotkeyListener.toggleTeleportHelper(player.uniqueId)
-
-        player.sendText {
-            appendSuccessPrefix()
-            if (newState) {
-                success("Der Teleport-Helfer wurde aktiviert.")
-            } else {
-                success("Der Teleport-Helfer wurde deaktiviert.")
-            }
-        }
-    }
-}
